Improve checks for child status
wait() just waits for the first random child. That may not be correct. Instead, use waitpid() to wait for a very specific particular child.

Also, fix the logic inversion in the next if statement: we failed the test if the child did not exit normally *or* the exit status is nonzero, rather than "and" there.