Fix "slotlist" test
It turns out that if you don't ask for the number of available slots before testing against it, that you'll always end up having no slots... d'oh.

Reorder things slightly so that this is no longer the case, and fix the bug that was exposed in the test by doing that.