Don't try to connect to the card before it's necessary
Currently, when we do a C_GetSlotList() with the "with token" parameter
set to TRUE, we read all the data from the card, even though we can't do
much with that data yet. This is a waste of CPU cycles, since we do it
again later on, and also makes things unnecessarily slow when we don't
actually need the card.

Add a parameter to the status update functions that states whether we
are interested in whether the card is knownw, and don't try to update
card statuses if not.