Fix failure to find things
std::string::find_first_of does *not* return -1, it returns
std::string::npos. While this often happens to be the same as (long)-1
and cast to unsigned, relying on that relies on undefined behaviour, and
aggressively optimizing compilers may therefore optimize the whole if()
structure away.

Fix by using the correct values.