Add a more limited version of readdata
The readdata test reads the whole card and dumps the data to stdout
(either in string form or in hex). This works with applet 1.7 (and
below) cards, but not with 1.8 with 3 PIN codes.

Add a version of readdata which only reads data from the carddata
object, which will never require a PIN code. While at it, add an
assertion that ensures the PKCS#11 module doesn't accidentally return
data which doesn't belong to the requested object.