Rehash PIN handling
PIN handling was done by reading the P15 files to figure out which PINs
were available, and how to use them.

The problem with that is that a) P15 files don't contain the PIN_readEF;
and b) it takes time to read the P15 file from the card, when we can
know (based on the GET CARD DATA output) which PIN codes are available.

So, just drop all the PIN code parsing, and just hardcode the PINs, only
varying them based on GET CARD DATA

TODO:
- tPinObjective is repeated, once in C++ and once in C. Clean that up.
- CKU_CONTEXT_SPECIFIC handling.