Partially revert "Enable C_WaitForSlotEvent() everywhere"
Originally, we implemented C_WaitForSlotEvent() on Windows only, not on
Linux or macOS. This was because it caused issues under Firefox.

We can no longer reproduce those Firefox issues, which is good. However,
it currently appears that C_WaitForSlotEvent() results in
CKR_DEVICE_ERROR on macOS, which is bad.

For now, disable C_WaitForSlotEvent(). We can fix it up good later.