Implement a proper abstraction layer
I originally implemented a version of the serial port communication
based on the older code that was deliberately not portable. However,
this was pretty broken and did not entirely work, partially due to
having added an abstraction at the wrong location.

Implement a proper abstraction layer that *only* deals with serial port
communication, and rewrite the older code in terms of that.

TODO: port to Windows and macOS. Up next, once this has been validated
to work.