"More" does not mean "All"
Modern OpenSSL has "openssl rehash" for hashing trust directories. Older
versions of OpenSSL have "c_rehash" instead. Previously we thought that
"openssl rehash" was available in most places, but now it turns out that
this is not the case. The result is that the packages we ship on some
older distributions (at least Ubuntu 16.04, probably some of the CentOS
versions too, haven't checked everything) ship with an unhashed trust
directory, resulting in verification of cards' certificates failing.

Unfortunately we didn't notice this because the 'openssl' tool will just
produce help output and exit with an exit state of 0 (i.e., indicating
no problem) when "openssl rehash" is not available.

Fix by explicitly calling 'openssl help' and grepping the output for the
rehash command, and falling back to c_rehash if it's not available.

Fixes: 074027040e7da36fd18f48065e54158b43770f24
Ticket: INC0327830
(cherry picked from commit c0b185ab511a18e0dd23a4493169697a26c131e2)