Improve ordering of libraries
Ubuntu defaults to linking with --as-needed, which means that the
compiler will drop libraries that are not needed by the program being
linked.

This however only works incrementally; that is, if a library being
linked is not used by any of the objects before it on the command line,
then the library is dropped, even if later objects on the command line
do use it.

Since eject_card only interacts with the robot, and not with the PKCS#11
library, this means that it will fail to link, since libtestlib wants to
link to the PKCS#11 library but we just dropped that.

Fix by placing libtestlib first, so that the linker sees the required
symbols and doesn't drop the PKCS#11 library