Don't block on checking if data exists
WaitCommEvents() does NOT allow for checking whether data is waiting
without blocking in the case it is not; it is solely meant to sit and
wait for data to appear, unless non-blocking I/O is used (which Windows
calls "overlapping I/O" for some reason).

Instead, one needs to use the ClearCommError() function (!) and then
check the values of the cbInQue member of the struct to which a pointer
is passed as the third argument.

This of course throws away any and all communications errors that might
have occurred, but meh, can't be helped.