Don't ask to insert readers in *every* find_slot()
find_slot() requires a reader to be inserted, so we need to insert one
if we don't find a reader. We did this previously by unconditionally
telling the robot to insert the reader every time, but that has the
downside that we annoy the interactive user by asking the same thing
every time in case of the manual "robot", and that we make things wait
for a few seconds so that everything can settle in case of actual
robots.

Fix by checking whether a (possibly empty) slot can be found, before
asking the user to insert a reader.