Fix test suite
The test suite fails when C_WaitForSlotEvent issues an event after
C_Initialize. This is how the Linux implementation currently behaves;
however, it is not according to spec. Therefore, the test suite should
*not* fail on that behaviour.

Fix that. This should cause the Windows test to succeed, and the Linux
test to fail.

Up next: disable the C_WaitForSlotEvent function on Linux again.