Disable this again
Unfortunately, C_WaitForSlotEvent actually still has a few bugs under
Linux. We were no longer able to reproduce the problem under Firefox
that originally caused us to disable the implementation there, but
(unfortunately) that does not mean that the bug is gone.

As such, disable this for now. At least we know the core problem now,
and the test suite will test for that; so we should be able to figure
out why this is failing and fix it properly, at our leasure.