Don't fail on spurious "event found" messages
C_WaitForSlotEvent() is supposed to say CKR_NO_EVENT at some point, but
doesn't, at least not always. That's not perfect, but it won't cause
terrible issues, so keep that check disabled for now.

We might want to fix it at some point in the future though, so keep it
in but commented out.