Don't fail on spurious events
Sometimes we get an event when there really isn't an event necessary.
That's a bit annoying, but it isn't a showstopper; software will just
not find anything to do when it's told there's something.

Since no harm results from doing that, don't try to be more holy than
the pope, and allow the test suite to succeed.