Add explicit casts
Coverity says that we have buffer overflows here, because we do a memset
based on a struct address, and the memset is much larger than the size
of the struct.

While that is a false positive, there are also compilers who will
optimize based on struct sizes, so we don't want that.

Add an explicit cast to avoid this issue.

Coverity CID#192848 and 192847