Don't stat when we're going to rename later
This loop ran stat() or wstat() over a filename. If the stat failed
(most likely because the file did not exist), then we exit the loop, as
an optimization. Later on, we then do a rename.

While this works, the downside of this method is that it introduces a
time of check/time of use race, where the file may be removed the
rename.

This isn't critical for us (it's an optimization, after all), but it's
not so nice -- and Coverity complains about it.

Since rename actually will set errno to ENOENT if the old file does not
exist, it makes sense to just drop the stat and exit the loop when
rename fails.

Coverity CID#179263