Add assertion to satisfy scan-build
Scan-build, the clang static analyser, found some control flow whereby
we would have a nonzero number of slots returned by C_GetSlotList but we
wouldn't have allocated them, because CKR_BUFFER_TOO_SMALL wouldn't have
been returned.

There is an implicit assumption in the code that that will never happen,
because we'd then have CKR_BUFFER_TOO_SMALL and things would have been
reallocated. However, this assumption is never made explicit, and so
scan-build can't see it.

Since it would be a pretty terrible bug to not return
CKR_BUFFER_TOO_SMALL with count set to nonzero, it's a good idea to test
for that explicitly.