Rework CI config
We want to do a few things different.

First, we want to parametrize the builds in such a way that we need to
make modifications in only one location when changing supported
distributions.

Second, we want to do that in such a way that the "signing" repository
can import the file, too. This requires that it be in a separate file.

Third, we want to move the signing repository to using submodules, and
to rebuilding things by itself, rather than taking them from the regular
repository. This means we no longer need the "extra:" jobs, so drop
them.

Finally, rename all the jobs so their name starts with the stage they
run in, which makes things slightly easier later on.