Improve if condition
The ${VAR:-val} syntax in Korn-style shells (including bash) will assign
the value "val" to the variable $VAR unless that variable is already set
to some other value.

This means that at the position where this if condition was entered, the
EIDMW_SIGN_BUILD variable will never be empty (it would be set by the
previous line), and the then clause was therefore dead code

Having that message is useful, so update the if condition so that it
will trigger if the variable is set to 1, in whatever way it happened.