Use the correct "echo" form.
"echo" is a shell builtin in bash, which makes things go slightly faster
and has a few extra features over /usr/bin/echo.

"Echo" is not. It also isn't an existing command, except that the
default macOS file system is case insensitive -- so that means that it
resolves to the external /usr/bin/echo binary on most mac systems.

However, it is possible to install macOS with a case-sensitive
filesystem (which is useful for some of the extra features such a
filesystem would have). On such a system, the Echo command would not
exist and the build would break.

Fix by switching all the echo commands to lower case.