Remove dead code
Struct members whose values are the same for all cards (and aren't even
accessed from anywhere) can be dropped. Update the one place that
accesses csPath, and hardcode the exact value for the necessary file
there.

TODO: verify that the SelectFile is even necessary still.

Additionally, the PrKDF parser is now no longer used, so junk it, too.

While at it, also add a "keyType" member, which we set to RSA everywhere
for now, but which we can fill up with EC for elliptic curve keys later.