Stop using deprecated APIs
RHEL 6 (and derivatives) did not ship with Gtk3, so we had to retain
compatibility with Gtk2.

We've recently dropped support for RHEL6. Additionally, Gtk4 is on the
horizon, which drops *all* the deprecated APIs in Gtk3.

As such, drop their usage here, so that we can upgrade to Gtk4 when the
time comes.