Drop "license" key in about dialog.
Gtk3 (and above) support a "license-type" key, which contains everything
needed for LGPL 3. As such, drop the explicit stuff that we included
only for Gtk2 compatibility.

As a bonus, this will also make it use license-type (rather than
license), when using Gtk4 (because the macros didn't work out correctly
for Gtk4)