Better Gtk4 compatibility
GtkBuilder in Gtk4 does away with the <packing> thing.

Glade will add a packing node even when all values are set to default,
so we can drop all of them, except one: the "expand: true" property on
the tree view.

Set the vexpand property instead, which is the Gtk4-supported
alternative for that.