Fix printing of foreigner card on macOS
The "work_permit_mention" field was listed as "work_permit_type" instead on the ForeignerPrintView, which resulted in it not being found when a foreigner card was attempted to be printed.

Additionally, the "get_foreigner_fields()" function returned a few fields that actually weren't on the foreigner card (such as "xml_file_version", a meta field), resulting in things breaking down when trying to print from a file.

Fix both.