Clean the OpenSSL/libxml2 builds before starting a release build
Debug builds, as done in Xcode, disable the universal binary part, which
means you end up with a native-arch-only build for your project. That's
fine for a debug build, but if you try to do a release build later, then
xcodebuild doesn't notice that the static libraries which we compile
don't contain the not-native architecture and fails to rebuild them.

Fix by forcing a "make clean" before we enter the main build, which will
remove the static library so that the build will later rebuild the whole
thing.