Allow debugging eID viewer against raw card data
Sometimes it's helpful to be able to test functionality of the eID
Viewer against raw dumps of the data on some cards that helpful citizens
provide us.

Don't include that functionality at all time though, only if it has been
explicitly compiled in.

This reuses the fuzzing code, so add an extra parameter to the fuzzing
option that compiles in the fuzzing code but doesn't instrument for any
of the fuzzers that we currently support.