Add missing explicit cast
Older versions of GTK defined GHFunc as a function that returned a
gboolean. Our code still assumes that, and returns TRUE, even though the
most recent versions of GTK don't expect that anymore, which throws a
warning.

Since returning something that isn't used is not a major issue, just add
an explicit cast to silence the warning.