Add missing callback
The GETPIN operation in libassuan requires that a callback is passed to
place the entered PIN code. We did that for the "get current PIN"
operation, but somehow missed adding that for the "get new PIN"
operation, which meant that the new PIN was never found, and in fact
libassuan returned an error that it didn't have anything to pass the
data to.

Fix in the obvious way.