Fix tags
"tags" are not merely random thingies; they are used to select which
runner to use. We need to use the system that has the robot connected to
it, otherwise we can't access the card reader.

As there is no runner with a "test" tag, don't try to look for it.