Stop using deprecated APIs to load dynamic module plugins
NSAddImage has been deprecated since macOS 10.5(!), with the replacement being dlopen().

We have a dlopen() implementation which we use for *some* builds, but apparently not all. Since NSAddImage does not have any way of signalling that the module could not be loaded, it crashes the application on error.

This is obviously not a good idea.

Make the USING_DL_OPEN #define a no-op, and always use the dlopen() implementation everywhere except on Windows.