Don't allow so many failures
While we were working on getting this up and running, it made sense to
not stop the whole build when one minor subthing fails. This is no
longer a good idea.

Do allow builds to fail that test things which we don't officially
support (i.e., gtk4, FreeBSD, and Debian sid), or things that are
currently known to be broken (while it is being fixed), but remove the
"allow_failure: true" from everything else.