Refactor settings into a preferences window
We have a number of preferences so far, which are all in one menu item
or another.

This will end up very confusing in the end, so move all the preferences
into a custom dialog box and drop the (now) superfluous menu items.

While at it, add a preference item for deciding whether to check for
updates at startup