Run the perl script on a Linux container, not on macOS
We wanted to avoid doing the artifact dance, but unfortunately the perl
installation on macOS does not come with the XML::SimpleObject module,
so we would have to install it there, and then maintain it, etc. That's
too involved.

Instead, do the artifact dance and then do the actual tests inside a
Linux container.