Drop confusing and unecessary log message
Originally, we weren't sure whether signature files were supposed to be
the lenght of the signature.

It was clarified later on that certain signature files will always be
the maximum possible length of a signature (which may end up being up to
4 bytes more than the actual length of the signature) to allow for
rekeying of the card.

As such, this is an expected message, so there is no need to log a
warning about that.

Closes: gh-162