Depend on pinentry-x11, rather than pinentry
pinentry is satisfiable by "pinentry-curses", which will not work with
the most likely use case of the eID middleware (running it as part of a
graphical desktop).

The curses interface is also available with other pinentry
implementations, so if there is someone who really wants to use a
non-graphical version then they can still do so.

Signed-off-by: Wouter Verhelst <wouter.verhelst@zetes.com>