Conflict with the OpenSC PKCS#11 module
The OpenSC module claims it supports the Belgian eID card, but it only
supports applet 1.7, not 1.8. The result is that users who have OpenSC
installed may or may not successfully authenticate, depending on whether
their browser prefers OpenSC over the eID software (or not).

To avoid this situation, we can conflict with OpenSC. Users who need
OpenSC for other things will hate us for that, but that can't be helped.