Make the opensc-pkcs11 dependency versioned
OpenSC is about to release their next version containing our patch that
fixes their bug, and they'll call it 0.26.0. So declare that we don't
conflict with versions containing the required fix.

While here, also use the Breaks: relationship for .deb distributions
rather than the Conflicts: one, which is a better (more modern) way of
expressing that.