Update the way of searching for OpenSSL
OpenSSL used to have a "exec_prefix" pkg-config variable that allowed us
to figure out where the "openssl" binary could be found.

As of OpenSSL 3.3, this pkg-config variable no longer exists, so we
can't use it anymore.

Migrate to AC_PATH_PROG to avoid that issue.

While here, remove the "c_rehash" fallback that we used to require for
older distributions, as none of the distributions that require it are
supported anymore.