Update detection of assuan
The old AM_PATH_LIBASSUAN used to pull in gpg-error, too, but that
apparently is no longer the case, which makes the build fail if
configure was generated on Debian trixie or later.

Update to use PKG_CHECK_MODULES instead.

In case the PKG_CHECK_MODULES macro installs some initialization code
the first time it is called, as this is a conditional call that is not
always invoked, move the pcsc-lite check up (which is unconditional), so
that we don't condition out that potential initialization code