Backcompat with older libassuans
libassuan.pc is installed on recent distributions, but not (yet) on EL8.

AM_PATH_LIBASSUAN works on those, but not on more recent systems that
require an explicit libgpg-error to be added.

So, try libassuan.pc (and gpg-error.pc) first, and fall back on
AM_PATH_LIBASSUAN if we can't find that.