Debian builds: use unshare mode
Modern sbuild uses unshare rather than chroot, and builds the
build environment on demand (if it doesn't exist yet), removing the need
for us to do this manually.

Apart from that, the tool we used to build a chroot calls debootstrap
behind the scenes, and something in there did not mount /proc on a
chroot before calling apt, which causes long builds and timeouts.
According to Julian (apt developer):

< juliank> wouter: you don't have proc mounted
< juliank> And high file limit  14:30
< juliank> It closes all FDs up to the ulimit

...so avoid all that by using the modern way of building things.