deb-sbuild: base on ubuntu
Building packages using the unshare mode of sbuild and the mmdebstrap
stuff makes things easier, but also means that we need to configure
more.

Using the ubuntu version of sbuild means it comes with preconfiguration
for Ubuntu (which we do need) that makes things just a tad bit easier,
so do that for now.